Oprava horneho menu: - korektne zobrazenie pri jednotlivych velkostiach - skratenie mena, ak je prilis dlhe
Dokoncenie administratorskej sekcie:
- admin nemoze zmazat sam seba
- admin nemoze menit rolu samemu sebe
- admin moze menit rolu ostatnym pouzivatelom

+++ App/Views/Admin/index.view.php
@@ -1,6 +1,7 @@
 <?php
 use App\Models\User;
 
+/** @var \App\Core\IAuthenticator $auth */
 /** @var User[] $data */
 ?>
 
@@ -9,7 +10,7 @@ use App\Models\User;
 
     <div class="card pouzivatel" style="width: 30rem;">
         <div class="card-body">
-            <div class="container px-4 text-center">
+            <div class="container-fluid px-4 text-center">
                 <div class="row g-0 text-center">
                     <div class="col-sm-6 col-md-8">
                         <div class="p-3">
@@ -18,7 +19,9 @@ use App\Models\User;
                     </div>
                     <div class="col-6 col-md-4">
                         <div class="p-3">
-                            <a href="?c=admin&a=delete&id=<?php echo $user->getId() ?>" class="deleteUser card-link">Vymazat</a>
+                            <?php if ($user->getId() != $auth->getLoggedUserId()) { ?>
+                                <a href="?c=admin&a=delete&id=<?php echo $user->getId() ?>" class="deleteUser card-link">Vymazat</a>
+                            <?php } ?>
                         </div>
                     </div>
                 </div>
@@ -33,26 +36,32 @@ use App\Models\User;
         </ul>
 
         <div class="container overflow-hidden text-center">
-            <div class="row gx-5">
-                <div class="col">
-                    <div class="p-3">
-                        <select class="form-select" aria-label="vyber role">
+            <form method="post" action="?c=admin&a=modify">
+                <?php /** @var \App\Models\User $data */
+                if ($user->getId()) { ?>
+                    <input type="hidden" name="id" value="<?php echo $user->getId() ?>">
+                <?php } ?>
+
+                <div>
+                    <div class="inline-block-child">
+                        <select class="form-select" name="rola" aria-label="Default select example">
                             <option selected>Zmena role</option>
-                            <option value="1">Pouzivatel: u</option>
-                            <option value="2">Administrator: a</option>
+                            <option value="u">Pouzivatel: u</option>
+                            <option value="a">Administrator: a</option>
                         </select>
                     </div>
-                </div>
-                <div class="col">
-                    <div class="p-3">
-                        <a href="?c=admin&modify&id=<?php echo $user->getId() ?>" class="loginTlacidlo btn btn-outline-dark col btn-md btn-block" role="button">Potvrdit</a>
+
+                    <div class="inline-block-child">
+                        <?php if ($user->getId() != $auth->getLoggedUserId()) { ?>
+                            <input type="submit" value="Potvrdit" class="submitTlacidlo btn btn-outline-dark col btn-md btn-block" role="button">
+                        <?php } else { ?>
+                            <input disabled type="submit" value="Potvrdit" class="submitTlacidlo btn btn-outline-dark col btn-md btn-block" role="button">
+                        <?php } ?>
                     </div>
                 </div>
-            </div>
+            </form>
         </div>
 
     </div>
 
 <?php } ?>
-
-<div class="medzeraVelka"></div>
